fix : 패키지 구조 변경 및 dto 이름 수정
- auth 관련 dto를 gamul/domain/dto에서  gamul/api/auth/domain/dto로 위치 변경
- MemberRequestDto, MemberResponseDto -> SignUpRequestDto, SignUpResponseDto로 이름 변경
