+++ src/main/java/com/gamul/gamul/api/auth/domain/dto/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.gamul.gamul.domain.dto;
+package com.gamul.gamul.api.auth.domain.dto;
 
 import com.gamul.gamul.domain.entity.Authority;
 import com.gamul.gamul.domain.entity.Member;
@@ -14,23 +14,14 @@ import javax.validation.constraints.NotBlank;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class MemberRequestDto {
-
-    @NotBlank
-    private String name;
-
+public class LoginRequestDto {
     @Email
     private String email;
 
     @NotBlank
     private String password;
 
-    public Member toMember(PasswordEncoder passwordEncoder) {
-        return Member.builder()
-                .name(name)
-                .email(email)
-                .password(passwordEncoder.encode(password))
-                .authority(Authority.ROLE_USER)
-                .build();
+    public UsernamePasswordAuthenticationToken toAuthentication(){
+        return new UsernamePasswordAuthenticationToken(email, password);
     }
 }
