+++ src/main/java/com/gamul/gamul/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.gamul.gamul.service;
 
-import com.gamul.gamul.domain.dto.MemberResponseDto;
+import com.gamul.gamul.api.auth.domain.dto.SignUpResponseDto;
 import com.gamul.gamul.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,9 +12,9 @@ public class MemberService {
     private final MemberRepository memberRepository;
 
     @Transactional(readOnly = true)
-    public MemberResponseDto getMemberInfo(String email) {
+    public SignUpResponseDto getMemberInfo(String email) {
         return memberRepository.findByEmail(email)
-                .map(MemberResponseDto::of)
+                .map(SignUpResponseDto::of)
                 .orElseThrow(()->new RuntimeException("유저 정보가 없습니다."));
     }
 }
